requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
joblib==1.2.0
numpy==1.23.4
pandas==1.5.0
pickle5==0.0.11
python-dateutil==2.8.2
pytz==2022.5
scikit-learn==1.1.2
scipy==1.9.2
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
